web/oxxovision/cloud-functions/analyze-image/requirements.txt
functions-framework==3.3.0
Pillow==9.5.0
requests==2.31.0 